Extract the BillingPeriod to a separate file, which can be used both by a ViewModel and a repository
